+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.57.0
+
+* Add a `split($string, $separator, $limit: null)` function to `sass:string`
+  that splits a string into separate substrings based on a separator string.
+
 ## 1.56.2
 
 ### Embedded Sass